+++ src/Entities/Users/NewUserAddress.php
@@ -9,6 +9,10 @@ class NewUserAddress extends NamedCoordinatedAddress implements JsonSerializable
      * @var string
      */
     private $note;
+    /**
+     * @var bool
+     */
+    private $favourite;
 
     /**
      * @param string $name
@@ -21,8 +25,9 @@ class NewUserAddress extends NamedCoordinatedAddress implements JsonSerializable
      * @param string $note
      * @param string $latitude
      * @param string $longitude
+     * @param bool   $favourite
      */
-    public function __construct($name, $streetNumber, $streetName, $complex, $suburb, $city, $postalCode, $note, $latitude, $longitude)
+    public function __construct($name, $streetNumber, $streetName, $complex, $suburb, $city, $postalCode, $note, $latitude, $longitude, $favourite = false)
     {
         parent::__construct(
             $streetNumber,
@@ -36,6 +41,7 @@ class NewUserAddress extends NamedCoordinatedAddress implements JsonSerializable
             $longitude
         );
         $this->note = $note;
+        $this->favourite = $favourite;
     }
 
     /**
@@ -47,6 +53,14 @@ class NewUserAddress extends NamedCoordinatedAddress implements JsonSerializable
     }
 
     /**
+     * @return boolean
+     */
+    public function isFavourite()
+    {
+        return $this->favourite;
+    }
+
+    /**
      * Specify data which should be serialized to JSON
      */
     public function jsonSerialize()
@@ -54,6 +68,7 @@ class NewUserAddress extends NamedCoordinatedAddress implements JsonSerializable
         $json = parent::jsonSerialize();
 
         $json['note'] = $this->getNote();
+        $json['favourite'] = $this->isFavourite();
 
         return $json;
     }